Cleans the code.
